Extracted QnAMaker logic from corresponding middleware
