+++ libraries/Microsoft.Bot.Builder.Ai/QnAMakerMiddleware.cs
@@ -1,104 +1,21 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using System;
 using System.Linq;
 using System.Net.Http;
 using System.Threading.Tasks;
 using Microsoft.Bot.Builder.Middleware;
 using Microsoft.Bot.Schema;
-using Newtonsoft.Json;
 
 namespace Microsoft.Bot.Builder.Ai
 {
-
-    public class QueryResult
-    {
-        [JsonProperty("questions")]
-        public string[] Questions { get; set; }
-
-        [JsonProperty("answer")]
-        public string Answer { get; set; }
-
-        [JsonProperty("score")]
-        public float Score { get; set; }
-    }
-
-    public class QueryResults
-    {
-        [JsonProperty("answers")]
-        public QueryResult[] Answers { get; set; }
-    }
-
-
-    public class QnAMakerOptions
-    {
-        public string SubscriptionKey { get; set; }
-        public string KnowledgeBaseId { get; set; }
-        public float ScoreThreshold { get; set; }
-        public int Top { get; set; }
-    }
-
     public class QnAMakerMiddleware : Middleware.IReceiveActivity
     {
-        public const string qnaMakerServiceEndpoint = "https://westus.api.cognitive.microsoft.com/qnamaker/v2.0/knowledgebases/";
-        private readonly string _answerUrl;
-        private readonly QnAMakerOptions _options;
-        private readonly HttpClient _httpClient;
-
-        public const string APIManagementHeader = "Ocp-Apim-Subscription-Key";
-        public const string JsonMimeType = "application/json"; 
+        private readonly QnAMaker _qnaMaker;
 
         public QnAMakerMiddleware(QnAMakerOptions options, HttpClient httpClient)
         {
-
-            _options = options ?? throw new ArgumentNullException(nameof(options));
-            this._httpClient = httpClient ?? throw new ArgumentNullException(nameof(httpClient));
-
-            if (string.IsNullOrEmpty(options.KnowledgeBaseId))
-            {
-                throw new ArgumentException(nameof(options.KnowledgeBaseId));
-            }
-
-            this._answerUrl = $"{qnaMakerServiceEndpoint}{options.KnowledgeBaseId}/generateanswer";
-
-            if (_options.ScoreThreshold == 0)
-            {
-                _options.ScoreThreshold = 0.3F;
-            }
-
-            if (_options.Top == 0)
-            {
-                _options.Top = 1;
-            }                        
-        }
-
-        public async Task<QueryResult[]> GetAnswers(string question)
-        {
-            var request = new HttpRequestMessage(HttpMethod.Post, this._answerUrl);
-            
-            string jsonRequest = JsonConvert.SerializeObject(new
-            {
-                question,
-                top = this._options.Top
-            }, Formatting.None);
-
-            var content = new StringContent(jsonRequest, System.Text.Encoding.UTF8, JsonMimeType);
-            content.Headers.Add(APIManagementHeader, this._options.SubscriptionKey);
-
-            var response = await this._httpClient.PostAsync(this._answerUrl, content).ConfigureAwait(false);
-            if (response.IsSuccessStatusCode)
-            {
-                var jsonResponse = await response.Content.ReadAsStringAsync().ConfigureAwait(false);
-                var results = JsonConvert.DeserializeObject<QueryResults>(jsonResponse);
-                foreach (var answer in results.Answers)
-                {
-                    answer.Score = answer.Score / 100;
-                }
-
-                return results.Answers.Where(answer => answer.Score > this._options.ScoreThreshold).ToArray();
-            }
-            return null;
+            _qnaMaker = new QnAMaker(options, httpClient);
         }
 
         public async Task ReceiveActivity(IBotContext context, MiddlewareSet.NextDelegate next)
@@ -108,7 +25,7 @@ namespace Microsoft.Bot.Builder.Ai
                 var messageActivity = context.Request.AsMessageActivity();
                 if (!string.IsNullOrEmpty(messageActivity.Text))
                 {
-                    var results = await this.GetAnswers(messageActivity.Text.Trim()).ConfigureAwait(false);
+                    var results = await _qnaMaker.GetAnswers(messageActivity.Text.Trim()).ConfigureAwait(false);
                     if (results.Any())
                     {
                         context.Reply(results.First().Answer);
@@ -116,7 +33,7 @@ namespace Microsoft.Bot.Builder.Ai
                 }
             }
 
-            await next().ConfigureAwait(false); 
+            await next().ConfigureAwait(false);
         }
     }
 }
